컴포넌트별로 파일화하여 가독성 증가함
