+++ src/pages/junsobKim/Main/Nav_kim.scss
@@ -1,0 +1,67 @@
+#mainNav {
+  display: flex;
+  top: 0;
+  flex-direction: row;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  height: 50px;
+  margin-bottom: 20px;
+  border-bottom: 0.5px solid rgb(128, 128, 128);
+  background-color: #fafafa;
+  gap: 200px;
+
+  #logos {
+    display: flex;
+    align-items: center;
+    width: 150px;
+    height: 30px;
+
+    .westarLogo {
+      display: inline-block;
+      width: 25px;
+      height: 25px;
+      background-image: url("https://cdn-icons-png.flaticon.com/512/1384/1384031.png");
+      background-position: center;
+      background-size: cover;
+    }
+
+    .westarLogo2 {
+      margin-left: 10px;
+      padding-left: 10px;
+      border-left: 1px solid rgb(128, 128, 128);
+      color: black;
+      font-family: "Lobster";
+      font-size: 20px;
+      text-decoration: none;
+    }
+  }
+
+  #mainProfileThings {
+    display: flex;
+    align-items: center;
+    width: 150px;
+    height: 30px;
+    gap: 20px;
+
+    .profileThings {
+      display: inline-block;
+      width: 25px;
+      height: 25px;
+      background-position: center;
+      background-size: cover;
+    }
+
+    #mainCompass {
+      background-image: url("https://s3.ap-northeast-2.amazonaws.com/cdn.wecode.co.kr/bearu/explore.png");
+    }
+
+    #mainNavHeart {
+      background-image: url("https://s3.ap-northeast-2.amazonaws.com/cdn.wecode.co.kr/bearu/heart.png");
+    }
+
+    #mainMyPage {
+      background-image: url("https://s3.ap-northeast-2.amazonaws.com/cdn.wecode.co.kr/bearu/profile.png");
+    }
+  }
+}
